requirements.txt
# 接口测试框架依赖包

# ============== 核心测试框架 ==============
pytest>=7.0.0
pytest-html>=3.1.0
pytest-xdist>=3.0.0
allure-pytest>=2.13.0

# ============== HTTP请求与API测试 ==============
requests>=2.28.0

# ============== 配置文件处理 ==============
PyYAML>=6.0
python-dotenv>=1.0.0

# ============== 数据处理 ==============
jsonpath>=0.82
pandas>=2.0.0
deepmerge>=1.1.0

# ============== 文档处理 ==============
python-docx>=0.8.11
PyPDF2>=3.0.0

# ============== AI SDK ==============
zhipuai>=2.1.0

# ============== 开发工具 ==============
black>=23.0.0
flake8>=6.0.0
